requirements.txt
click==6.6
Flask==0.11.1
pywebview==1.3
